refactor: split implementation
